backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async HTTP Client
httpx>=0.24.0,<0.27.0
aiohttp>=3.9.1

# Database
supabase>=2.3.0
postgrest>=0.13.0

# AI/ML APIs
google-generativeai==0.3.2
pillow==10.2.0

# Data Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Rate Limiting
slowapi==0.1.9

# Date/Time
python-dateutil==2.8.2

